+++ mitra_cli/src/commands/profile.rs
@@ -56,7 +56,6 @@ impl ListUnreachableActors {
 
 /// Delete user
 #[derive(Parser)]
-#[command(visible_alias = "delete-account", alias = "delete-profile")]
 pub struct DeleteUser {
     id_or_name: String,
 }