+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added option to generate `Like` activities to `create-activity` command.
 - Added `federation.no_proxy` configuration parameter.
+- Added command groups: `account`, `ap`, `config`, `emoji`, `filter`, `invite`, `media`.
 
 ### Changed